Integerfunctions (#101)
* Moved Ackermann Function to new class
* Removed unused variables
* Renamed intdiv to fullIntdiv
* Added indiv and mod functions.
