+++ src/Number/ArbitraryInteger.php
@@ -99,8 +99,6 @@ class ArbitraryInteger implements ObjectArithmetic
             }
             // Convert to base 256
             $base256 = new ArbitraryInteger(0);
-            $base_obj = new ArbitraryInteger($base);
-            $place_value = new ArbitraryInteger(1);
             $length = strlen($number);
             for ($i = 0; $i < $length; $i++) {
                 $chr = ord($number[$i]);
@@ -195,6 +193,61 @@ class ArbitraryInteger implements ObjectArithmetic
         // Strip leading chr(0) entries.
         $this->base256 = $value;
         $this->positive = $positive;
+    }
+
+    /**************************************************************************
+     * UNARY FUNCTIONS
+     **************************************************************************/
+
+    /**
+     * Multiply by -1
+     *
+     * If $this is zero, then do nothing
+     */
+    public function negate(): ArbitraryInteger
+    {
+        $result = new ArbitraryInteger(0);
+        $base256 = $this->base256;
+        $result->setVariables($base256, $base256 == chr(0) ? true : !$this->positive);
+        return $result;
+    }
+
+    /**
+     * Integer Square Root
+     *
+     * Calculate the largest integer less than the square root of an integer.
+     * https://en.wikipedia.org/wiki/Integer_square_root
+     */
+    public function isqrt(): ArbitraryInteger
+    {
+        $length = strlen($this->base256);
+        // Start close to the value, at a number around half the digits.
+        $X = self::fromBinary(substr($this->base256, 0, intdiv($length, 2) + 1), true);
+        $lastX = $X;
+        $converge = false;
+        while (!$converge) {
+            $NX = $this->intdiv($X);
+            $X = $X->add($NX)->intdiv(2);
+            if ($X->equals($lastX) || $X->equals($lastX->add(1))) {
+                $converge = true;
+            } else {
+                $lastX = $X;
+            }
+        }
+        return $lastX;
+    }
+
+    /**
+     * Absolute Value
+     *
+     * @return ArbitraryInteger
+     */
+    public function abs(): ArbitraryInteger
+    {
+        $result = new ArbitraryInteger(0);
+        $base256 = $this->base256;
+        $result->setVariables($base256, true);
+        return $result;
     }
 
     /**************************************************************************
@@ -338,7 +391,7 @@ class ArbitraryInteger implements ObjectArithmetic
     /**
      * Integer Division
      */
-    public function intdiv($divisor): array
+    public function fullIntdiv($divisor): array
     {
         if ($this->lessThan($divisor)) {
             return [new ArbitraryInteger(0), $this];
@@ -394,59 +447,16 @@ class ArbitraryInteger implements ObjectArithmetic
         return [$int, $mod];
     }
 
-    /**************************************************************************
-     * UNARY FUNCTIONS
-     **************************************************************************/
-
-    /**
-     * Multiply by -1
-     *
-     * If $this is zero, then do nothing
-     */
-    public function negate(): ArbitraryInteger
+    public function mod($divisor): ArbitraryInteger
     {
-        $result = new ArbitraryInteger(0);
-        $base256 = $this->base256;
-        $result->setVariables($base256, $base256 == chr(0) ? true : !$this->positive);
-        return $result;
+        list($int, $mod) = $this->fullIntdiv($divisor);
+        return $mod;
     }
 
-    /**
-     * Integer Square Root
-     *
-     * Calculate the largest integer less than the square root of an integer.
-     * https://en.wikipedia.org/wiki/Integer_square_root
-     */
-    public function isqrt(): ArbitraryInteger
+    public function intdiv($divisor): ArbitraryInteger
     {
-        $length = strlen($this->base256);
-        // Start close to the value, at a number around half the digits.
-        $X = self::fromBinary(substr($this->base256, 0, intdiv($length, 2) + 1), true);
-        $lastX = $X;
-        $converge = false;
-        while (!$converge) {
-            list($NX, $mod) = $this->intdiv($X);
-            list($X, $mod) = $X->add($NX)->intdiv(2);
-            if ($X->equals($lastX) || $X->equals($lastX->add(1))) {
-                $converge = true;
-            } else {
-                $lastX = $X;
-            }
-        }
-        return $lastX;
-    }
-
-    /**
-     * Absolute Value
-     *
-     * @return ArbitraryInteger
-     */
-    public function abs(): ArbitraryInteger
-    {
-        $result = new ArbitraryInteger(0);
-        $base256 = $this->base256;
-        $result->setVariables($base256, true);
-        return $result;
+        list($int, $mod) = $this->fullIntdiv($divisor);
+        return $int;
     }
 
     /**
@@ -470,37 +480,6 @@ class ArbitraryInteger implements ObjectArithmetic
         return $result;
     }
 
-    /**
-     * Ackermann Function
-     * A well known highly recursive function which produces very large numbers
-     *
-     * https://en.wikipedia.org/wiki/Ackermann_function
-     *
-     * @param $m
-     * @param $n
-     * @return ArbitraryInteger
-     */
-    public static function ackermann($m, $n): ArbitraryInteger
-    {
-        $m = self::prepareParameter($m);
-        $n = self::prepareParameter($n);
-        if ($m->equals(0)) {
-            return $n->add(1);
-        } elseif ($m->equals(1)) {
-            return $n->add(2);
-        } elseif ($m->equals(2)) {
-            return $n->leftShift(1)->add(3);
-        } elseif ($m->equals(3)) {
-            $one = new ArbitraryInteger(1);
-            // 2^(n+3) - 3
-            return $one->leftShift($n->add(3))->subtract(3);
-        } elseif ($n->equals(0)) {
-            return ArbitraryInteger::ackermann($m->subtract(1), 1);
-        } else {
-            return ArbitraryInteger::ackermann($m->subtract(1), ArbitraryInteger::ackermann($m, $n->subtract(1)));
-        }
-    }
-
     /**************************************************************************
      * BITWISE OPERATIONS
      **************************************************************************/
@@ -513,7 +492,7 @@ class ArbitraryInteger implements ObjectArithmetic
         $bits = self::prepareParameter($bits);
         $shifted_string = "";
         $length = strlen($this->base256);
-        list($bytes, $bits) = $bits->intdiv(8);
+        list($bytes, $bits) = $bits->fullIntdiv(8);
         $bits = $bits->toInteger();
         $carry = 0;
         for ($i = 0; $i < $length; $i++) {
